backend/requirements.txt
fastapi
uvicorn
pymongo
python-multipart
pycryptodome
cryptography
